move a few functions around to help with compiling
